+++ src/Oro/Bundle/UserBundle/Resources/config/services.yml
@@ -6,7 +6,7 @@ parameters:
     oro_user.security.provider.class: Oro\Bundle\UserBundle\Security\UserProvider
     oro_user.security.login.class:    Oro\Bundle\UserBundle\EventListener\LoginSubscriber
     oro_user.role.listener.class:     Oro\Bundle\UserBundle\EventListener\RoleListener
-    oro_user.entity_with_image.subscriber.class: Oro\Bundle\UserBundle\Entity\EventListener\UploadedImageSubscriber
+    oro_user.entity_with_image.subscriber.class: Pim\Bundle\UserBundle\EventSubscriber\UploadedImageSubscriber
     oro_user.change_password.subscriber.class:   Oro\Bundle\UserBundle\Form\EventListener\ChangePasswordSubscriber
 
 services:
